dev_requirements.txt
-r requirements.txt
django-extensions
werkzeug
django-debug-toolbar